@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>    
-    <TargetFramework>netcoreapp2.0</TargetFramework>    
+    <TargetFramework>netcoreapp3.1</TargetFramework>    
     <RootNamespace>Tutorial.Tests</RootNamespace>    
     <Authors />    
     <Company>Dixin Yan</Company>    
