@@ -1,7 +1,6 @@
 ﻿namespace Tutorial.Tests.LinqToEntities
 {
-    using System.Data.SqlClient;
-
+    using Microsoft.Data.SqlClient;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Tutorial.LinqToEntities;
 
