@@ -2,8 +2,8 @@
 {
     using System;
     using System.Data.Common;
-    using System.Data.SqlClient;
 
+    using Microsoft.Data.SqlClient;
     using Microsoft.EntityFrameworkCore;
     using Tutorial.LinqToEntities;
 
